Add testing framework

+++ src/simulator.hpp
@@ -125,7 +125,7 @@ private:
     std::unordered_map<OpCode,
         std::function<State(Instruction)>, OpCodeHash> m_inst_funcs;
     // Instruction called counter
-    std::unordered_map<OpCode, int32_t, OpCodeHash> m_inst_cnt;
+    std::unordered_map<OpCode, int64_t, OpCodeHash> m_inst_cnt;
     void initInstruction();
 
     static OpCode decodeOpCode(Instruction);
